Wed, Sep 7, 2022, 9:31 PM +00:00

+++ readability/readability.c
@@ -28,9 +28,9 @@ int main(void)
             sentences++;
         }
     }
-    float L = letters / words * 100
-    float S = sentences / words * 100
+    float L = (float) letters / (float) words * 100
+    float S = (float) sentences / (float) words * 100
 
     int index = round (0.0588 * L - 0.296 * S - 15.8);
-
+    
 }
